商品削除機能_修正１

+++ app/controllers/items_controller.rb
@@ -39,8 +39,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.user.id == current_user.id
+      item.destroy
+      redirect_to root_path
+    else
+      render :show
+    end
   end
 
   private
